@@ -1,13 +1,13 @@
 package com.tristanruecker.interviewexampleproject.models.objects.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
 import java.time.Instant;
 
 @MappedSuperclass
@@ -16,13 +16,13 @@ public abstract class AbstractDatabaseObject {
 
     @CreationTimestamp
     @Column(name = "creation_date")
-    @ApiModelProperty(hidden = true)
+    @Schema(hidden = true)
     @JsonIgnore
     private Instant creationDate;
 
     @UpdateTimestamp
     @Column(name = "modification_date")
-    @ApiModelProperty(hidden = true)
+    @Schema(hidden = true)
     @JsonIgnore
     private Instant modificationDate;
 
